Fix Queen being able to jump over enemy pieces

+++ sources/engine/pieces/Piece.cpp
@@ -16,7 +16,7 @@ bool    Piece::isMoveLegal(ChessEngine *engine, int8_t x, int8_t y) {
     if (!isMoveOnBoard(x, y))
         return false;
 
-    Piece *p = engine->getPieceFromPos({pos.x + x, pos.y + y});
+    Piece *p = engine->getPieceFromPos({(int8_t)(pos.x + x), (int8_t)(pos.y + y)});
     if (p)
         return canCapture(p);
